+++ docs/CHANGELOG.md
@@ -53,6 +53,8 @@ Version 1.9.0 (not released yet)
 - Add mod name to main menu
 - Make value of `spectate_mode_minimal_ui` persist between game launches
 - Add `version` command
+- Add `gore_level` command
+- Add gibbing at gore level 2 when enemies die from explosives
 
 [@is-this-c](https://github.com/is-this-c)
 - Support `©` in TrueType fonts